+++ src/Trarizon.Library.Collections/HISTORY.md
@@ -4,3 +4,4 @@
 
 - Reimplement `PrefixTree(Dictionary)` with node, the original implementation is renamed to `ContiguousPrefixTree(Dicitionary)` and marked as experimental
 - Add `TraAlgorithm.LevenshteinDistance`
+- Add `IEnumerable.Intersperse`